+++ src/Interfaces/address/Address.ts
@@ -1,0 +1,15 @@
+export default interface Address {
+    id?: string;
+    firstname?: string;
+    lastname?: string;
+    address?: string;
+    addressTwo?: string;
+    addressthree?: string;
+    city?: string;
+    company?: string;
+    zip?: string;
+    shipping?: boolean;
+    billing?: boolean;
+    default?: boolean;
+    phoneNumber?: string;
+}