+++ src/Interfaces/checkout/Checkout.ts
@@ -1,0 +1,4 @@
+export default interface Checkout {
+    canCheckout?: boolean;
+    message?: string;
+}