+++ src/Interfaces/subscription/Subscription.ts
@@ -1,4 +1,7 @@
 import { SubscriptionPricingPlan } from './SubscriptionPricingPlan';
+import SubscriptionTier from './SubscriptionTier';
+
+export type SubscriptionIntervalLength = 1 | 6 | 12;
 
 export interface APISubscription {
     id?: string;
@@ -13,6 +16,8 @@ export interface APISubscription {
     active?: number;
     created_at?: Date;
     updated_at?: Date;
+    subscription_pricing_plans?: Array<SubscriptionPricingPlan>;
+    subscription_tiers?: Array<SubscriptionTier>;
 }
 
 export interface Subscription {
@@ -28,5 +33,6 @@ export interface Subscription {
     active?: number;
     createdAt?: Date;
     updatedAt?: Date;
-    pricingPlans?: Array<SubscriptionPricingPlan>;
+    subscriptionPricingPlans?: Array<SubscriptionPricingPlan>;
+    subscriptionTiers?: Array<SubscriptionTier>;
 }