+++ src/Models/address/Address.ts
@@ -1,0 +1,7 @@
+import AddressInterface from '../../Interfaces/address/Address';
+
+export default class Address {
+    constructor(parameters: AddressInterface) {
+        Object.assign(this, parameters);
+    }
+}